Added filtering to dashboard sidebar (#74)
The top of the sidebar should have a dropdown with the following four
options:
1. Sponsorship Funds
2. Funding Items
3. Personal Purchase Requests
4. UW Finance Purchase Requests

Each option should have a checkbox that is checkable. The checked status
of each option will determine if that option's ticket type is displayed
in the sidebar. (i.e. if personal purchases and uw finance purchases are
the only options checked, then only those two types of tickets should be
displayed in the sidebar)

The tickets (filtered by option) should still be sorted by the latest
date updated, so the newest updated ticket will be at the top card shown
in the sidebar.