User info display (#76)
Created a user endpoint that fetches all users in the firebase project
using ``firebase-admin``. The [ticket
](https://github.com/WATonomous/wato_finance_system/issues/61) will
require us to add a local json file to our "backend" folder in order to
fetch user data. The instructions in README.md will explain the steps.

Fetched the user data to be able to display a hoverable component as
shown below:




![image](https://user-images.githubusercontent.com/83772958/212343846-13811d3d-79a5-42c0-91c9-69d6f7a1b0b9.png)